+++ src/accounts/nexus.ts
@@ -179,10 +179,11 @@ async function getPackedSignature(
   hash: Hex,
   validator: ValidatorConfig,
   transformSignature: (signature: Hex) => Hex = (signature) => signature,
+  defaultValidatorAddress: Address = NEXUS_DEFAULT_VALIDATOR_ADDRESS,
 ) {
   const signature = await signFn(hash)
   const validatorAddress =
-    validator.address === NEXUS_DEFAULT_VALIDATOR_ADDRESS
+    validator.address === defaultValidatorAddress
       ? zeroAddress
       : validator.address
   const packedSig = encodePacked(
@@ -198,6 +199,7 @@ async function getSmartAccount(
   owners: OwnerSet,
   validatorAddress: Address,
   sign: (hash: Hex) => Promise<Hex>,
+  defaultValidatorAddress: Address = NEXUS_DEFAULT_VALIDATOR_ADDRESS,
 ) {
   return getBaseSmartAccount(
     address,
@@ -207,6 +209,7 @@ async function getSmartAccount(
       return getMockSignature(owners)
     },
     sign,
+    defaultValidatorAddress,
   )
 }
 
@@ -217,6 +220,7 @@ async function getSessionSmartAccount(
   validatorAddress: Address,
   enableData: EnableSessionData | null,
   sign: (hash: Hex) => Promise<Hex>,
+  defaultValidatorAddress: Address = NEXUS_DEFAULT_VALIDATOR_ADDRESS,
 ) {
   return await getBaseSmartAccount(
     address,
@@ -254,6 +258,7 @@ async function getSessionSmartAccount(
         signature,
       )
     },
+    defaultValidatorAddress,
   )
 }
 
@@ -263,6 +268,7 @@ async function getGuardianSmartAccount(
   guardians: OwnerSet,
   validatorAddress: Address,
   sign: (hash: Hex) => Promise<Hex>,
+  defaultValidatorAddress: Address = NEXUS_DEFAULT_VALIDATOR_ADDRESS,
 ) {
   return await getBaseSmartAccount(
     address,
@@ -274,6 +280,7 @@ async function getGuardianSmartAccount(
     async (hash) => {
       return await sign(hash)
     },
+    defaultValidatorAddress,
   )
 }
 
@@ -283,6 +290,7 @@ async function getBaseSmartAccount(
   nonceValidatorAddress: Address,
   getStubSignature: () => Promise<Hex>,
   signUserOperation: (hash: Hex) => Promise<Hex>,
+  defaultValidatorAddress: Address,
 ): Promise<SmartAccount<SmartAccountImplementation<Abi, '0.7'>>> {
   return await toSmartAccount({
     client,
@@ -313,7 +321,7 @@ async function getBaseSmartAccount(
     },
     async getNonce(args) {
       const validatorAddress =
-        nonceValidatorAddress === NEXUS_DEFAULT_VALIDATOR_ADDRESS
+        nonceValidatorAddress === defaultValidatorAddress
           ? zeroAddress
           : nonceValidatorAddress
       const TIMESTAMP_ADJUSTMENT = 16777215n // max value for size 3