+++ src/orchestrator/types.ts
@@ -9,8 +9,8 @@ import type {
   optimism,
   optimismSepolia,
   polygon,
-  polygonAmoy,
   sepolia,
+  soneium,
   zksync,
 } from 'viem/chains'
 
@@ -19,7 +19,6 @@ type SupportedTestnet =
   | typeof baseSepolia.id
   | typeof arbitrumSepolia.id
   | typeof optimismSepolia.id
-  | typeof polygonAmoy.id
 type SupportedMainnet =
   | typeof mainnet.id
   | typeof base.id
@@ -27,7 +26,11 @@ type SupportedMainnet =
   | typeof optimism.id
   | typeof polygon.id
   | typeof zksync.id
-type SupportedOPStackMainnet = typeof optimism.id | typeof base.id
+  | typeof soneium.id
+type SupportedOPStackMainnet =
+  | typeof optimism.id
+  | typeof base.id
+  | typeof soneium.id
 type SupportedOPStackTestnet = typeof optimismSepolia.id | typeof baseSepolia.id
 type SupportedChain = SupportedMainnet | SupportedTestnet
 type SupportedTokenSymbol = 'ETH' | 'WETH' | 'USDC' | 'USDT'